+++ download_list.php
@@ -14,6 +14,7 @@ $url = "https://portal.fedsfm.ru/account/login";
 $options = array (
     CURLOPT_COOKIEJAR => __DIR__ . DIRECTORY_SEPARATOR . 'cookie.txt'
 );
+
 $output = CurlRequest::sendPostRequest($url, $post_data, $options);
 
 $obj = json_decode($output);
@@ -28,11 +29,41 @@ $options = array (
     CURLOPT_COOKIEFILE => __DIR__ . DIRECTORY_SEPARATOR . 'cookie.txt'
 );
 
+$directoryPath = __DIR__ . DIRECTORY_SEPARATOR . "activeSked";
+
+$needDownload = false;
+if (!file_exists($directoryPath)) {
+    $needDownload = true;
+} else {
+    $needDownload = zeroDownloadCount($options);
+}
+
+if (!$needDownload) {
+    return;
+}
+
 $url = "https://portal.fedsfm.ru/SkedDownload/GetActiveSked?type=dbf";
 $fileName = __DIR__ . DIRECTORY_SEPARATOR . "activeSked.zip";
 CurlRequest::downloadFile($url, $fileName, $options);
 
-$directoryPath = __DIR__ . DIRECTORY_SEPARATOR . "activeSked";
+
 FileOperation::deleteDirectory($directoryPath);
 FileOperation::extractZip($fileName, $directoryPath);
+
+
+
+function zeroDownloadCount($options) {
+    $url = "https://portal.fedsfm.ru/StartPage/TerroristCatalogUserDashboard";
+    $post_data = array ('rowIndex' => 0, 'pageLength' => 10);
+    $output = CurlRequest::sendPostRequest($url, $post_data, $options);
+    $obj = json_decode($output);
+    $data = $obj->data;
+    if ($data != null && is_array($data) && count($data) > 0) {
+        $downloadCount = (int) $data[0]->DownloadCount;
+        if ($downloadCount == 0) {
+            return true;
+        }
+    }
+    return false;
+}
